backend/requirements.txt
docopt
keyring
keyrings.alt
pycrypto
selenium
python-dateutil
qifparse
flask
werkzeug
flask_cors
flask_sqlalchemy
git+https://github.com/julianogouveia/simple-crypt.git
sqlalchemy
regex
yapf
unify